Basic multicamera working. Mouse look broken
