+++ GTAV_CustomCamera.csproj
@@ -50,9 +50,13 @@
     <Reference Include="System.Xaml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Camera\CustomCamera.cs" />
+    <Compile Include="Camera\LegacyCamera.cs" />
+    <Compile Include="Camera\RacingCamera.cs" />
     <Compile Include="Helper\InputR.cs" />
     <Compile Include="Helper\MathR.cs" />
-    <Compile Include="Scripts\CustomCameraV.cs" />
+    <Compile Include="Helper\Time.cs" />
+    <Compile Include="Script\CustomCameraV.cs" />
     <Compile Include="Easing\Easing.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Transform\Transform.cs" />
