+++ Helper/Time.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using CustomCameraVScript;
+using GTA.Native;
+
+namespace CustomCameraVScript
+{
+    public static class Time
+    {
+        public static float getDeltaTime()
+        {
+            return Function.Call<float>(Hash.TIMESTEP);
+        }
+    }
+}
